Fix compiler warning on 64bit
src/ct_driver.c: In function ‘chipsPreInitHiQV’:
src/ct_driver.c:1598: warning: format ‘%lX’ expects type ‘long unsigned int’, but argument 4 has type ‘CARD32’

+++ src/ct_driver.c
@@ -1595,7 +1595,7 @@ chipsPreInitHiQV(ScrnInfoPtr pScrn, int flags)
 #endif
 	    cPtr->IOAddress = cPtr->FbAddress + 0x400000L;
  	xf86DrvMsg(pScrn->scrnIndex, X_DEFAULT,
- 		   "IOAddress is set at 0x%lX.\n",cPtr->IOAddress);
+		   "IOAddress is set at 0x%lX.\n",(unsigned long)cPtr->IOAddress);
 	
     } else
 	xf86DrvMsg(pScrn->scrnIndex, from,
